Simplified Cache Lookup of GJChronologies.
This commit replaces the cache lookup using only the zone and then
iterating over the ArrayList donating possible candidates with a single
cache access. For this purpose a temporary cache key is instantiated.

In the cache hit case this code will produce a temporary GJCacheKey
object and then return the GJChronology, as compared to previous, where
the code was forced synchronised, an ArrayList iteration was performed
and an Instant was checked for equality.